requirements.txt
# Telegram Bot
python-telegram-bot==20.7

# Web scraping
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.16.0

# Image processing & AI
# Install IOPaint with all plugins (RealESRGAN, interactive seg, etc.)
iopaint[all_plugins]==1.6.0

# OpenAI for content generation
openai==1.54.0

# Environment variables
python-dotenv==1.0.0

# RunPod serverless
runpod==1.7.0
